+++ apps/template/src/index.tsx
@@ -5,7 +5,6 @@
 
 import React from "react"
 import { createRoot } from "react-dom/client"
-import "tailwindcss/tailwind.css"
 import App from "./App"
 
 const container = document.getElementById("root") as HTMLDivElement