+++ packages/ui-components/src/components/LoadingIndicator/LoadingIndicator.stories.tsx
@@ -32,6 +32,6 @@ export const SizedLoadingIndicator: Story = {
 
 export const ColoredLoadingIndicator: Story = {
   args: {
-    color: "jn-text-theme-info",
+    color: "jn:text-theme-info",
   },
 }