+++ packages/ui-components/src/components/TooltipTrigger/TooltipTrigger.component.tsx
@@ -44,7 +44,7 @@ export const TooltipTrigger = React.forwardRef<HTMLElement, TooltipTriggerProps>
       ...props,
       ...childrenProps,
       "data-state": state.open ? "open" : "closed",
-      className: childrenProps.className + `${state.disabled && " jn-cursor-default"}`,
+      className: childrenProps.className + `${state.disabled && " jn:cursor-default"}`,
     }
     const combinedProps = state.getReferenceProps(referencedProps)
 
@@ -57,7 +57,7 @@ export const TooltipTrigger = React.forwardRef<HTMLElement, TooltipTriggerProps>
       // The user can style the trigger based on the state
       data-state={state.open ? "open" : "closed"}
       {...state.getReferenceProps(props)}
-      className={`${className} ${state.disabled && " jn-cursor-default"}`}
+      className={`${className} ${state.disabled && " jn:cursor-default"}`}
     >
       {children}
     </button>