+++ packages/ui-components/src/dummyComponents/DummyComponent.tsx
@@ -13,16 +13,16 @@ const DummyComponent = ({ label }: DummyComponentProps) => {
   return (
     <div
       className="
-        jn-flex
-        jn-bg-juno-blue
-        jn-rounded
-        jn-text-white
-        jn-px-8
-        jn-py-3
-        jn-border
-        jn-border-juno-blue-7"
+        jn:flex
+        jn:bg-juno-blue
+        jn:rounded
+        jn:text-white
+        jn:px-8
+        jn:py-3
+        jn:border
+        jn:border-juno-blue-7"
     >
-      <div className="jn-m-auto">{label}</div>
+      <div className="jn:m-auto">{label}</div>
     </div>
   )
 }